requirements.txt
Flask==2.3.4
flask-pymongo==2.3.0
Flask-Bcrypt==1.0.1
Flask-JWT-Extended==4.5.1
python-dotenv==1.0.0
requests==2.31.0
email-validator==2.1.1
